@@ -1,4 +1,4 @@
-const { PrismaClient } = require('../generated/prisma');
+import { PrismaClient } from '../generated/prisma';
 const prisma = new PrismaClient();
 
 async function main() {
