@@ -4,34 +4,14 @@ import {
   OnModuleDestroy,
   OnModuleInit,
 } from '@nestjs/common';
-import path from 'path';
-
-import type { PrismaClient as PrismaClientType } from '../../../../generated/prisma/client';
-
-const getPrismaClient = (): typeof PrismaClientType => {
-  const root = process.cwd();
-  const clientPath = path.join(root, 'generated', 'prisma', 'client');
-
-  // eslint-disable-next-line @typescript-eslint/no-require-imports
-  const required = require(clientPath);
-
-  return (required?.PrismaClient ??
-    required?.default ??
-    required) as typeof PrismaClientType;
-};
-
-const BasePrismaClient = getPrismaClient();
+import { PrismaClient } from 'generated/prisma';
 
 @Injectable()
 export class PrismaService
-  extends BasePrismaClient
+  extends PrismaClient
   implements OnModuleInit, OnModuleDestroy
 {
   private readonly logger = new Logger(PrismaService.name);
-
-  constructor() {
-    super();
-  }
 
   async onModuleInit() {
     try {
