@@ -7,17 +7,22 @@ import { TelemetriesService } from './telemetries.service';
 export class TelemetriesController {
   constructor(private readonly telemetriesService: TelemetriesService) {}
 
-  @Get('temperature')
-  async getTemperature(
-    @Query() query: TelemetryQueryDto,
-  ): Promise<TelemetryPointDto[]> {
-    const deviceKey = query.deviceKey;
+  private parseMinutes(query: TelemetryQueryDto): number {
     const rawMinutes = (query as unknown as Record<string, any>)['minutes'];
     const minutes =
       typeof rawMinutes === 'number' ? rawMinutes : Number(rawMinutes);
     if (!Number.isFinite(minutes)) {
       throw new BadRequestException('Minutes must be a number');
     }
+    return minutes;
+  }
+
+  @Get('temperature')
+  async getTemperature(
+    @Query() query: TelemetryQueryDto,
+  ): Promise<TelemetryPointDto[]> {
+    const deviceKey = query.deviceKey;
+    const minutes = this.parseMinutes(query);
     return this.telemetriesService.getTemperatureSeries(deviceKey, minutes);
   }
 
@@ -26,12 +31,43 @@ export class TelemetriesController {
     @Query() query: TelemetryQueryDto,
   ): Promise<TelemetryPointDto[]> {
     const deviceKey = query.deviceKey;
-    const rawMinutes = (query as unknown as Record<string, any>)['minutes'];
-    const minutes =
-      typeof rawMinutes === 'number' ? rawMinutes : Number(rawMinutes);
-    if (!Number.isFinite(minutes)) {
-      throw new BadRequestException('Minutes must be a number');
-    }
+    const minutes = this.parseMinutes(query);
     return this.telemetriesService.getHumiditySeries(deviceKey, minutes);
+  }
+
+  @Get('light')
+  async getLight(
+    @Query() query: TelemetryQueryDto,
+  ): Promise<TelemetryPointDto[]> {
+    const deviceKey = query.deviceKey;
+    const minutes = this.parseMinutes(query);
+    return this.telemetriesService.getLightSeries(deviceKey, minutes);
+  }
+
+  @Get('rain')
+  async getRain(
+    @Query() query: TelemetryQueryDto,
+  ): Promise<TelemetryPointDto[]> {
+    const deviceKey = query.deviceKey;
+    const minutes = this.parseMinutes(query);
+    return this.telemetriesService.getRainSeries(deviceKey, minutes);
+  }
+
+  @Get('servo')
+  async getServo(
+    @Query() query: TelemetryQueryDto,
+  ): Promise<TelemetryPointDto[]> {
+    const deviceKey = query.deviceKey;
+    const minutes = this.parseMinutes(query);
+    return this.telemetriesService.getServoSeries(deviceKey, minutes);
+  }
+
+  @Get('all')
+  async getAll(
+    @Query() query: TelemetryQueryDto,
+  ): Promise<Record<string, TelemetryPointDto[]>> {
+    const deviceKey = query.deviceKey;
+    const minutes = this.parseMinutes(query);
+    return this.telemetriesService.getAllSeries(deviceKey, minutes);
   }
 }
