@@ -7,7 +7,11 @@ export class TelemetriesService {
   constructor(private readonly prismaService: PrismaService) {}
 
   private readonly CAP = 200;
-  private readonly THRESHOLD_FACTOR = 20;
+
+  private roundToTwo(value: number) {
+    if (!Number.isFinite(value)) return value;
+    return Math.round(value * 100) / 100;
+  }
 
   private ensureMinutesBound(minutes: number) {
     if (!Number.isFinite(minutes) || minutes < 1) return 1;
@@ -22,89 +26,20 @@ export class TelemetriesService {
     return device?.id ?? null;
   }
 
-  private mapToTelemetryPoints(
-    rows: Array<{
-      createdAt: Date;
-      temperature?: number | null;
-      humidity?: number | null;
-    }>,
-    valueKey: 'temperature' | 'humidity',
-  ) {
-    return rows.map((r) => {
-      const value = valueKey === 'temperature' ? r.temperature : r.humidity;
-      return new TelemetryPointDto(r.createdAt, value ?? 0);
-    });
-  }
-
-  private async runAggregation(
-    deviceId: string,
-    whereStart: Date,
-    valueKey: 'temperature' | 'humidity',
-  ): Promise<TelemetryPointDto[] | null> {
-    const cap = this.CAP;
-    const field = valueKey === 'temperature' ? 'temperature' : 'humidity';
-
-    const pipeline = [
-      {
-        $match: {
-          deviceId: { $oid: deviceId },
-          createdAt: { $gte: whereStart },
-        },
-      },
-      { $sort: { createdAt: 1 } },
-      {
-        $bucketAuto: {
-          groupBy: '$createdAt',
-          buckets: cap,
-          output: {
-            avgValue: { $avg: `$${field}` },
-            avgTime: { $avg: { $toLong: '$createdAt' } },
-          },
-        },
-      },
-    ];
-
-    try {
-      const res = (await this.prismaService.$runCommandRaw({
-        aggregate: 'telemetries',
-        pipeline,
-        cursor: {},
-      })) as any;
-
-      const firstBatch = res?.cursor?.firstBatch ?? res;
-      if (!Array.isArray(firstBatch)) return null;
-
-      const points = firstBatch
-        .map((b: any) => {
-          const avgTime = b?.avgTime ?? null;
-          const avgValue = b?.avgValue ?? null;
-          if (avgTime == null || avgValue == null) return null;
-          const timeNum =
-            typeof avgTime === 'number' ? avgTime : Number(avgTime);
-          return new TelemetryPointDto(new Date(timeNum), Number(avgValue));
-        })
-        .filter(Boolean) as TelemetryPointDto[];
-
-      return points;
-    } catch (error) {
-      console.warn(
-        'DB aggregation failed, falling back to app sampling',
-        error,
-      );
-      return null;
-    }
-  }
-
-  private sampleTelemetryPoints(
-    rows: Array<{
-      createdAt: Date;
-      temperature?: number | null;
-      humidity?: number | null;
-    }>,
-    valueKey: 'temperature' | 'humidity',
+  private samplePointsGeneric(
+    rows: Array<{ createdAt: Date } & Record<string, any>>,
+    valueFn: (r: any) => number,
   ): TelemetryPointDto[] {
-    const cap = 200;
-    if (rows.length <= cap) return this.mapToTelemetryPoints(rows, valueKey);
+    const cap = this.CAP;
+    if (rows.length <= cap) {
+      return rows
+        .map((r) => {
+          const raw = valueFn(r);
+          const val = this.roundToTwo(raw);
+          return new TelemetryPointDto(r.createdAt, val);
+        })
+        .filter(Boolean);
+    }
 
     const bucketSize = rows.length / cap;
     const sampled: TelemetryPointDto[] = [];
@@ -119,7 +54,7 @@ export class TelemetriesService {
       let count = 0;
       let timeSum = 0;
       for (const item of bucket) {
-        const v = valueKey === 'temperature' ? item.temperature : item.humidity;
+        const v = valueFn(item);
         if (v === undefined || v === null) continue;
         sum += v;
         count++;
@@ -127,36 +62,30 @@ export class TelemetriesService {
       }
       if (count === 0) continue;
       const avgValue = sum / count;
+      const rounded = this.roundToTwo(avgValue);
       const avgTime = new Date(Math.floor(timeSum / count));
-      sampled.push(new TelemetryPointDto(avgTime, avgValue));
+      sampled.push(new TelemetryPointDto(avgTime, rounded));
     }
 
     return sampled;
   }
 
-  private async fetchAndSample(
+  private async fetchAndSampleGeneric(
     deviceId: string,
     whereStart: Date,
-    valueKey: 'temperature' | 'humidity',
+    select: Record<string, boolean>,
+    valueFn: (r: any) => number,
   ): Promise<TelemetryPointDto[]> {
-    const select =
-      valueKey === 'temperature'
-        ? { createdAt: true, temperature: true }
-        : { createdAt: true, humidity: true };
-
     const rows = await this.prismaService.telemetry.findMany({
       where: { deviceId, createdAt: { gte: whereStart } },
       orderBy: { createdAt: 'asc' },
       select,
     });
 
-    const normalized = rows as unknown as Array<{
-      createdAt: Date;
-      temperature?: number | null;
-      humidity?: number | null;
-    }>;
-
-    return this.sampleTelemetryPoints(normalized, valueKey);
+    const normalized = rows as unknown as Array<
+      { createdAt: Date } & Record<string, any>
+    >;
+    return this.samplePointsGeneric(normalized, valueFn);
   }
 
   async getTemperatureSeries(deviceKey: string, minutes: number) {
@@ -165,27 +94,14 @@ export class TelemetriesService {
     const bounded = this.ensureMinutesBound(minutes);
     const whereStart = new Date(Date.now() - bounded * 60 * 1000);
     const deviceId = await this.findDeviceIdByKey(deviceKey);
-    if (!deviceId) {
-      return [];
-    }
+    if (!deviceId) return [];
 
-    const cap = this.CAP;
-    const threshold = cap * this.THRESHOLD_FACTOR;
-
-    const total = await this.prismaService.telemetry.count({
-      where: { deviceId, createdAt: { gte: whereStart } },
-    });
-
-    if (total > threshold) {
-      const points = await this.runAggregation(
-        deviceId,
-        whereStart,
-        'temperature',
-      );
-      if (points && points.length > 0) return points;
-    }
-
-    return this.fetchAndSample(deviceId, whereStart, 'temperature');
+    return this.fetchAndSampleGeneric(
+      deviceId,
+      whereStart,
+      { createdAt: true, temperature: true },
+      (r) => r.temperature ?? 0,
+    );
   }
 
   async getHumiditySeries(deviceKey: string, minutes: number) {
@@ -194,26 +110,122 @@ export class TelemetriesService {
     const bounded = this.ensureMinutesBound(minutes);
     const whereStart = new Date(Date.now() - bounded * 60 * 1000);
     const deviceId = await this.findDeviceIdByKey(deviceKey);
-    if (!deviceId) {
-      return [];
-    }
+    if (!deviceId) return [];
 
-    const cap = this.CAP;
-    const threshold = cap * this.THRESHOLD_FACTOR;
+    return this.fetchAndSampleGeneric(
+      deviceId,
+      whereStart,
+      { createdAt: true, humidity: true },
+      (r) => r.humidity ?? 0,
+    );
+  }
 
-    const total = await this.prismaService.telemetry.count({
+  async getLightSeries(deviceKey: string, minutes: number) {
+    if (!deviceKey) throw new BadRequestException('deviceKey is required');
+
+    const bounded = this.ensureMinutesBound(minutes);
+    const whereStart = new Date(Date.now() - bounded * 60 * 1000);
+    const deviceId = await this.findDeviceIdByKey(deviceKey);
+    if (!deviceId) return [];
+
+    return this.fetchAndSampleGeneric(
+      deviceId,
+      whereStart,
+      { createdAt: true, lightIntensity: true },
+      (r) => r.lightIntensity ?? 0,
+    );
+  }
+
+  async getRainSeries(deviceKey: string, minutes: number) {
+    if (!deviceKey) throw new BadRequestException('deviceKey is required');
+
+    const bounded = this.ensureMinutesBound(minutes);
+    const whereStart = new Date(Date.now() - bounded * 60 * 1000);
+    const deviceId = await this.findDeviceIdByKey(deviceKey);
+    if (!deviceId) return [];
+
+    return this.fetchAndSampleGeneric(
+      deviceId,
+      whereStart,
+      { createdAt: true, rainStatus: true },
+      (r) => (r.rainStatus === 'RAIN' ? 1 : 0),
+    );
+  }
+
+  async getServoSeries(deviceKey: string, minutes: number) {
+    if (!deviceKey) throw new BadRequestException('deviceKey is required');
+
+    const bounded = this.ensureMinutesBound(minutes);
+    const whereStart = new Date(Date.now() - bounded * 60 * 1000);
+    const deviceId = await this.findDeviceIdByKey(deviceKey);
+    if (!deviceId) return [];
+
+    return this.fetchAndSampleGeneric(
+      deviceId,
+      whereStart,
+      { createdAt: true, servoStatus: true },
+      (r) => (r.servoStatus === 'OPEN' ? 1 : 0),
+    );
+  }
+
+  async getAllSeries(
+    deviceKey: string,
+    minutes: number,
+  ): Promise<{
+    temperature: TelemetryPointDto[];
+    humidity: TelemetryPointDto[];
+    light: TelemetryPointDto[];
+    rain: TelemetryPointDto[];
+    servo: TelemetryPointDto[];
+    mode: TelemetryPointDto[];
+  }> {
+    if (!deviceKey) throw new BadRequestException('deviceKey is required');
+
+    const bounded = this.ensureMinutesBound(minutes);
+    const whereStart = new Date(Date.now() - bounded * 60 * 1000);
+    const deviceId = await this.findDeviceIdByKey(deviceKey);
+    if (!deviceId)
+      return {
+        temperature: [],
+        humidity: [],
+        light: [],
+        rain: [],
+        servo: [],
+        mode: [],
+      };
+
+    const rows = await this.prismaService.telemetry.findMany({
       where: { deviceId, createdAt: { gte: whereStart } },
+      orderBy: { createdAt: 'asc' },
+      select: {
+        createdAt: true,
+        temperature: true,
+        humidity: true,
+        lightIntensity: true,
+        rainStatus: true,
+        servoStatus: true,
+        mode: true,
+      },
     });
 
-    if (total > threshold) {
-      const points = await this.runAggregation(
-        deviceId,
-        whereStart,
-        'humidity',
-      );
-      if (points && points.length > 0) return points;
-    }
+    const normalized = rows as unknown as Array<
+      { createdAt: Date } & Record<string, any>
+    >;
 
-    return this.fetchAndSample(deviceId, whereStart, 'humidity');
+    const tempFn = (r: any) => r.temperature ?? 0;
+    const humFn = (r: any) => r.humidity ?? 0;
+    const lightFn = (r: any) => r.lightIntensity ?? 0;
+    const rainFn = (r: any) => (r.rainStatus === 'RAIN' ? 1 : 0);
+    const servoFn = (r: any) => (r.servoStatus === 'OPEN' ? 1 : 0);
+    const modeFn = (r: any) => (r.mode === 'AUTO' ? 1 : 0);
+
+    const temperature = this.samplePointsGeneric(normalized, tempFn);
+    const humidity = this.samplePointsGeneric(normalized, humFn);
+    const light = this.samplePointsGeneric(normalized, lightFn);
+    const rain = this.samplePointsGeneric(normalized, rainFn);
+    const servo = this.samplePointsGeneric(normalized, servoFn);
+    const mode = this.samplePointsGeneric(normalized, modeFn);
+
+    return { temperature, humidity, light, rain, servo, mode };
   }
 }
